Further development of MRbyGP
1) Replacing constant from int to float and constant simplification : Almost Done
We finished the development of constant simplification with float variable. But because there could be exceptional cases which we don't expect, we are trying to debug simplification part with many generations of run.

2) Insertion of Power function : Done
We finished the development of power function and debugging part.

3) GA Refactoring with two subject trees(LHS, RHS): Developing
Because deap library generally assumes that the population is the list of "one individual",
We are trying to investigate some tricky ways for dealing with two individuals at once.

We first check the way of changing (tree, tree) tuple as a individual.
In this approach, we should change and even define Crossover and Mutate method of deap.
Therefore, we guess that it needs more time for development of this part.